# openapi/koru_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "koru",
    "description": "An API for managing long-term shared expenses.",
    "contact": {
      "name": "Raphael Biland",
      "email": "raphael.biland@outlook.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.4.0"
  },
  "paths": {
...
    "/groups/{group_id}/members": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Adds the user making the request to the group whose id is provided, given a valid token from `/groups/{group_id}/token`.",
        "description": "Adds the user making the request to the group whose id is provided, given a valid token from `/groups/{group_id}/token`.\n\nThis action can only be performed by the member himself.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExample (replace GROUP_ID ith the group UUID):\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"token\":\"TOKEN\",\"color\":{\"red\":0,\"green\":255,\"blue\":0}}' -b cookie2 \"http://localhost:8000/groups/GROUP_ID/members\"\n```\n",
        "operationId": "join_group",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGroupPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member joined successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "User is already a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Groups"
        ],
        "summary": "Changes the color of the member making the request, within the provided group id.",
        "description": "Changes the color of the member making the request, within the provided group id.\n\nThis action can only be performed by the member himself.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExample (replace GROUP_ID ith the group UUID):\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"color\":{\"red\":255,\"green\":255,\"blue\":255}}' -b cookie -X PATCH \"http://localhost:8000/groups/GROUP_ID/members\"\n```\n",
        "operationId": "change_color",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group Uuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeColorPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Color changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or group id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs in the user given valid credentials.",
        "description": "Logs in the user given valid credentials.\n\nIt will return a `Set-Cookie` header, that should be used in next authenticated requests\n\nExample:\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"password\":\"123\",\"email\":\"r@r1.com\"}' -c cookie \"http://localhost:8000/login\"\n```",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Auth cookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors in login request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeColorPayload": {
        "type": "object",
        "required": [
          "color"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          }
        }
      },
      "ColorDto": {
        "type": "object",
        "required": [
          "red",
          "green",
          "blue"
        ],
        "properties": {
          "blue": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "green": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "red": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          }
        }
      },
...
      "ErrorData": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ErrorData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "IdData": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "IdResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IdData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "JoinGroupPayload": {
        "type": "object",
        "required": [
          "token",
          "color"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "LoginPayload": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MessageData": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id"
      }
    }
  },
...
}